removed  import ray from top level impor
